画像や動画に対応し，テキストを返す関数を追加
